improve user views and add tests

+++ app/users/views.py
@@ -1,6 +1,4 @@
-from datetime import datetime
-
-from flask import Blueprint, flash, redirect, render_template, request, url_for
+from flask import Blueprint, abort, flash, redirect, render_template, request, url_for
 from werkzeug import Response
 
 from app.extensions import database
@@ -26,6 +24,9 @@ def view_user(user_id: int) -> str:
     A route that allows claims officer fillout a form for a particular user
     """
     user_data = User.query.get(user_id)
+    if not user_data:
+        abort(404)
+
     return render_template("users/user_data.html", user_data=user_data)
 
 
@@ -34,28 +35,20 @@ def edit_user(user_id: int) -> Response | str:
     """
     A route for editing a user record
     """
+    user_data = User.query.get(user_id)
+    if not user_data:
+        abort(404)
+
     form = AddUser()
 
-    user_data = User.query.get(user_id)
-
     if request.method == "POST":
-        name = request.form.get("name")
-        gender = request.form.get("gender")
-        salary = request.form.get("salary")
-        date_of_birth = request.form.get("date_of_birth")
-
-        dt_obj = datetime.strptime(date_of_birth, "%Y-%m-%d")
-
         if form.validate_on_submit():
-            user_data.name = name
-            user_data.gender = gender
-            user_data.salary = salary
-            user_data.date_of_birth = dt_obj
+            form.populate_obj(user_data)
 
             database.session.commit()
 
             flash("User updated successfully.", "success")
-            return redirect(url_for("users.edit_user", id=user_data.id))
+            return redirect(url_for("users.edit_user", user_id=user_data.id))
 
     return render_template("users/edit_user.html", user_data=user_data, form=form)
 
@@ -67,19 +60,11 @@ def add_user() -> Response | str:
 
     """
     form = AddUser()
-
     if request.method == "POST":
-        name = request.form.get("name")
-        gender = request.form.get("gender")
-        salary = request.form.get("salary")
-        date_of_birth = request.form.get("date_of_birth")
-
-        dt_obj = datetime.strptime(date_of_birth, "%Y-%m-%d")
-
         if form.validate_on_submit():
-            new_user = User(
-                name=name, gender=gender, salary=salary, date_of_birth=dt_obj
-            )
+            new_user = User()
+            form.populate_obj(new_user)
+
             database.session.add(new_user)
             database.session.commit()
 
@@ -88,9 +73,8 @@ def add_user() -> Response | str:
 
         flash("You entered and invalid form data", "danger")
         return render_template("users/create_user.html", form=form)
-    else:
 
-        return render_template("users/create_user.html", form=form)
+    return render_template("users/create_user.html", form=form)
 
 
 @users.route("/<int:user_id>/delete", methods=["POST"])
@@ -99,6 +83,9 @@ def delete_user(user_id: int) -> Response:
     A route for deleting a user
     """
     user_data = User.query.get(user_id)
+    if not user_data:
+        abort(404)
+
     database.session.delete(user_data)
     database.session.commit()
